server/requirements.txt
sqlalchemy==1.4.41
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
anthropic==0.3.0
pydantic>=1.9.0,<2.0.0